add link to selected item
